added part code of User-Based Collaborative Filtering Algorithm

+++ src/main/java/com/yuki/Utils/MatrixOperationsUtils.java
@@ -50,12 +50,14 @@ public class MatrixOperationsUtils {
         return result;
     }
 
-    public static double cosineSimilarity(int[] vectorA, int[] vectorB) {
+    public static Double cosineSimilarity(Double[] vectorA, Double[] vectorB) {
         double dotProduct = 0.0;
         double normA = 0.0;
         double normB = 0.0;
         for (int i = 0; i < vectorA.length; i++) {
+//            计算分子
             dotProduct += vectorA[i] * vectorB[i];
+//            计算分母
             normA += Math.pow(vectorA[i], 2);
             normB += Math.pow(vectorB[i], 2);
         }
